core: use 32-bit math to compute dir entry offset
DirEntryIndexToOffset() was using 64-bit multiplication to compute the
offset of a directory entry within a block.  Since the maximum supported
block size is much less than UINT32_MAX, and the index involved in the
computation is also relative to the start of the block, it is not
possible for this multiplication to overflow a 32-bit integer.

+++ core/driver/dir.c
@@ -1071,7 +1071,7 @@ static uint64_t DirEntryIndexToOffset(
     {
         ullOffset += NODEHEADER_SIZE;
     }
-    ullOffset += (uint64_t)ulDirentsIntoBlock * DIRENT_SIZE;
+    ullOffset += ulDirentsIntoBlock * DIRENT_SIZE;
 
     return ullOffset;
 }
